+++ src/nl/hannahsten/texifyidea/project/LatexProjectSdkSetupValidator.kt
@@ -14,8 +14,8 @@ import com.intellij.ui.EditorNotificationPanel
 import nl.hannahsten.texifyidea.LatexLanguage
 import nl.hannahsten.texifyidea.file.LatexFileType
 import nl.hannahsten.texifyidea.modules.LatexModuleType
-import nl.hannahsten.texifyidea.settings.LatexSdk
-import nl.hannahsten.texifyidea.settings.LatexSdkUtil
+import nl.hannahsten.texifyidea.settings.sdk.LatexSdk
+import nl.hannahsten.texifyidea.settings.sdk.LatexSdkUtil
 
 /**
  * https://jetbrains.org/intellij/sdk/docs/reference_guide/project_model/sdk.html#assisting-in-setting-up-an-sdk
