+++ src/nl/hannahsten/texifyidea/util/files/LatexPackageLocationCache.kt
@@ -1,7 +1,7 @@
 package nl.hannahsten.texifyidea.util.files
 
 import com.intellij.openapi.project.Project
-import nl.hannahsten.texifyidea.settings.LatexSdkUtil
+import nl.hannahsten.texifyidea.settings.sdk.LatexSdkUtil
 import java.io.BufferedReader
 import java.io.IOException
 import java.io.InputStreamReader
